Extract Property to a separate file
