+++ lib/grumlin/typing.rb
@@ -28,23 +28,6 @@ module Grumlin
       end
     end
 
-    class Property
-      attr_reader :key, :value
-
-      def initialize(value)
-        @key = value[:key]
-        @value = Typing.cast(value[:value])
-      end
-
-      def inspect
-        "p[#{key}->#{value}]"
-      end
-
-      def to_s
-        inspect
-      end
-    end
-
     CASTABLE_TYPES = [Hash, String, Integer, TrueClass, FalseClass].freeze
 
     class << self
